Moved main functions to MainMenu class
